refactor(api): C6 — devtypes/connections/internal-repos services; structure ratchet closes (ADR-0015)
- NEW api/devtypes_service.py (13 bodies: prompt-template + devtype-prompt
  CRUD, harnesses, Dev Type CRUD/rename/credentials, assignments),
  api/connections_service.py (secret-ref validators + _HARNESS_VAR_RE + 8
  bodies: secrets, harness-secrets, secrets-check, registry, pmo/forge
  tests), api/internal_repos_service.py (8 bodies: internal repos + skills).
- main.py: 1133 → 780 lines. Every endpoint is a forward with unchanged
  decorator/name/signature (AST-verified); singletons read from module
  globals at call time so app_main monkeypatches keep working. Bodies moved
  verbatim except the sanctioned reload_connections → reload param.
- RATCHET CLOSED: test_structure_guards.py gains
  test_main_route_bodies_stay_thin — every @app route body ≤4 statements
  (docstring excluded) with a documented residual allowlist (dispatch_hello,
  mapper/OAuth trio, small read-side runs/log/clear endpoints) that may
  shrink, never grow. AGENTS.md anti-pattern entry + docs/01 api tree.
- Test retargets: _HARNESS_VAR_RE import → connections_service;
  save_config/save_dev_type/delete_dev_type monkeypatches in the rename
  test → devtypes_service; comment pointer fixes in config.py/secrets.py.

Verified: ruff clean; full suite 638 passed / 1 skipped via pytest_app.sh
(+1 = the new guard). The main.py split (C4-C6) is complete.
