+++ AGENTS.md
@@ -97,6 +97,7 @@ Design and refactors must respect SOLID. Prefer **deep modules** (small interfac
 - Giant untested functions in `orchestrator.py` without a public-seam test
 - `except Exception: pass` that swallows real failures without logging — lint-enforced: ruff `BLE001`; every blanket catch is narrowed or carries `# noqa: BLE001 — <justification>` naming its contract (docs/15 §7)
 - Mutating production modules only “to make the test pass” by weakening invariants
+- New endpoint bodies in `api/main.py` or new attributes bound onto `MissionManager` after its class body — main.py is composition root + ≤4-statement route forwards, orchestrator behavior lives in module functions taking `mgr` (ADR-0015; enforced by `tests/test_structure_guards.py`)
 
 ## Docker images: Bake only
